[#165] Features in Mylyn that aren't in SimRel
Repackage `org.eclipse.egit.mylyn.ui` bundle and its test fragment to
`org.eclipse.mylyn.egit` namespace
