--- mylyn.egit/org.eclipse.egit.mylyn.ui/build.properties
@@ -1,8 +1,0 @@
-source.. = src/
-output.. = bin/
-bin.includes = META-INF/,\
-               .,\
-               plugin.xml,\
-               plugin.properties,\
-               about.html
-src.includes = about.html
