--- mylyn.egit/org.eclipse.egit.mylyn.ui/plugin.properties
@@ -1,15 +1,0 @@
-###############################################################################
-# Copyright (c) 2011 Chris Aniszczyk and others.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License 2.0
-# which accompanies this distribution, and is available at
-# https://www.eclipse.org/legal/epl-2.0/
-#
-# SPDX-License-Identifier: EPL-2.0
-#
-# Contributors:
-#      Chris Aniszczyk <caniszczyk@gmail.com> - initial API and implementation
-###############################################################################
-Bundle-Vendor = Eclipse Mylyn
-Bundle-Name = Mylyn EGit UI
-hyperlinkDetector.name = Git Commits
