--- mylyn.egit/org.eclipse.egit.mylyn.ui/plugin.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<?eclipse version="3.4"?>
-<plugin>
-   <extension
-         point="org.eclipse.egit.ui.commitMessageProvider">
-      <commitMessageProvider
-            class="org.eclipse.egit.internal.mylyn.ui.commit.MylynCommitMessageProvider">
-      </commitMessageProvider>
-   </extension>
-
-   <extension point="org.eclipse.core.runtime.adapters">
-    <factory adaptableType="org.eclipse.egit.ui.internal.history.SWTCommit"
-             class="org.eclipse.egit.internal.mylyn.ui.commit.TaskReferenceFactory">
-       <adapter type="org.eclipse.mylyn.team.ui.AbstractTaskReference"/>
-    </factory>
-	<factory adaptableType="org.eclipse.egit.ui.internal.synchronize.model.GitModelCommit"
-             class="org.eclipse.egit.internal.mylyn.ui.commit.TaskReferenceFactory">
-       <adapter type="org.eclipse.mylyn.team.ui.AbstractTaskReference"/>
-    </factory>
-   </extension>
-   <extension
-         point="org.eclipse.ui.workbench.texteditor.hyperlinkDetectors">
-      <hyperlinkDetector
-            activate="true"
-            class="org.eclipse.egit.internal.mylyn.ui.CommitHyperlinkDetector"
-            id="org.eclipse.egit.mylyn.ui.commitHyperlinkDetector"
-            name="%hyperlinkDetector.name"
-            targetId="org.eclipse.mylyn.tasks.ui.TaskEditor">
-      </hyperlinkDetector>
-   </extension>
-   <extension
-         point="org.eclipse.egit.ui.branchNameProvider">
-      <branchNameProvider
-            class="org.eclipse.egit.internal.mylyn.ui.commit.ActiveTaskBranchNameProvider">
-      </branchNameProvider>
-   </extension>
-   <extension
-         point="org.eclipse.mylyn.tasks.core.taskActivationListeners">
-      <listener
-            id="org.eclipse.egit.mylyn.ui.commit.CommitTextTaskActivationListener"
-            class="org.eclipse.egit.internal.mylyn.ui.commit.CommitTextTaskActivationListener">
-      </listener>
-   </extension>
-</plugin>
