--- mylyn.egit/org.eclipse.egit.mylyn.ui/pom.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-   Copyright (C) 2011, Chris Aniszczyk <caniszczyk@gmail.com>
-
-   All rights reserved. This program and the accompanying materials
-   are made available under the terms of the Eclipse Public License 2.0
-   which accompanies this distribution, and is available at
-   https://www.eclipse.org/legal/epl-2.0/
-
-   SPDX-License-Identifier: EPL-2.0
--->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.eclipse.mylyn</groupId>
-    <artifactId>org.eclipse.mylyn.parent</artifactId>
-    <version>3.26.1-SNAPSHOT</version>
-    <relativePath>../../mylyn.releng/org.eclipse.mylyn.parent</relativePath>
-  </parent>
-
-  <artifactId>org.eclipse.egit.mylyn.ui</artifactId>
-  <packaging>eclipse-plugin</packaging>
-
-  <name>EGit Mylyn UI</name>
-
- <build>
-    <plugins>
-      <plugin>
-        <groupId>org.eclipse.tycho</groupId>
-        <artifactId>tycho-source-plugin</artifactId>
-      </plugin>
-    </plugins>
-  </build>
-
-</project>
