--- mylyn.egit/org.eclipse.egit.mylyn.ui/src/org/eclipse/egit/internal/mylyn/ui/messages.properties
@@ -1,2 +1,0 @@
-CommitHyperlinkDetector_CommitNotFound=Commit not found
-CommitHyperlinkDetector_CommitNotFoundInRepositories=The commit {0} was not found in any of the configured repositories.
