+++ mylyn.egit/org.eclipse.mylyn.egit.feature/feature.xml
@@ -31,7 +31,7 @@
    </requires>
 
    <plugin
-         id="org.eclipse.egit.mylyn.ui"
+         id="org.eclipse.mylyn.egit.ui"
          download-size="0"
          install-size="0"
          version="0.0.0"
