+++ mylyn.egit/org.eclipse.mylyn.egit.ui.tests/.classpath
