+++ mylyn.egit/org.eclipse.mylyn.egit.ui.tests/.settings/org.eclipse.core.runtime.prefs
