+++ mylyn.egit/org.eclipse.mylyn.egit.ui.tests/build.properties
@@ -1,0 +1,17 @@
+###############################################################################
+# Copyright (c) 2023 Contributors to the Eclipse Foundation
+# 
+# This program and the accompanying materials are made available under the
+# terms of the Eclipse Public License 2.0 which is available at
+# https://www.eclipse.org/legal/epl-2.0/.
+# 
+# SPDX-License-Identifier: EPL-2.0
+# 
+# Contributors:
+#   See git history
+###############################################################################
+
+source.. = src/
+output.. = bin/
+bin.includes = META-INF/,\
+               .
