+++ mylyn.egit/org.eclipse.mylyn.egit.ui/.classpath
