+++ mylyn.egit/org.eclipse.mylyn.egit.ui/.settings/org.eclipse.core.runtime.prefs
