+++ mylyn.egit/org.eclipse.mylyn.egit.ui/.settings/org.eclipse.jdt.core.prefs
