+++ mylyn.egit/org.eclipse.mylyn.egit.ui/.settings/org.eclipse.mylyn.tasks.ui.prefs
