+++ mylyn.egit/org.eclipse.mylyn.egit.ui/.settings/org.eclipse.pde.api.tools.prefs
