+++ mylyn.egit/org.eclipse.mylyn.egit.ui/META-INF/MANIFEST.MF
@@ -1,0 +1,34 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: %Bundle-Name
+Automatic-Module-Name: org.eclipse.egit.mylyn.ui
+Bundle-SymbolicName: org.eclipse.mylyn.egit.ui;singleton:=true
+Bundle-Version: 3.26.1.qualifier
+Bundle-Vendor: %Bundle-Vendor
+Require-Bundle: org.eclipse.jface;bundle-version="[3.21.0,4.0.0)",
+ org.eclipse.core.resources;bundle-version="[3.13.800,4.0.0)",
+ org.eclipse.core.runtime;bundle-version="[3.19.0,4.0.0)",
+ org.eclipse.ui.workbench;bundle-version="[3.120.0,4.0.0)",
+ org.eclipse.mylyn.tasks.core;bundle-version="[3.20.0,4.0.0)",
+ org.eclipse.mylyn.tasks.ui;bundle-version="[3.20.0,4.0.0)",
+ org.eclipse.mylyn.team.ui;bundle-version="[3.20.0,4.0.0)",
+ org.eclipse.mylyn.context.core;bundle-version="[3.20.0,4.0.0)",
+ org.eclipse.jface.text;bundle-version="[3.16.400,4.0.0)",
+ org.eclipse.team.core;bundle-version="[3.8.1100,4.0.0)"
+Bundle-RequiredExecutionEnvironment: JavaSE-11
+Bundle-ActivationPolicy: lazy
+Bundle-Localization: plugin
+Export-Package: org.eclipse.mylyn.egit.internal.ui;version="3.26.1";x-friends:="org.eclipse.egit.ui.test",
+ org.eclipse.mylyn.egit.internal.ui.commit;version="3.26.1";x-internal:=true
+Import-Package: org.eclipse.egit.core;version="[6.2.0,7.0.0)",
+ org.eclipse.egit.core.internal;version="[6.2.0,7.0.0)",
+ org.eclipse.egit.core.synchronize;version="[6.2.0,7.0.0)",
+ org.eclipse.egit.ui;version="[6.2.0,7.0.0)",
+ org.eclipse.egit.ui.internal.commit;version="[6.2.0,7.0.0)",
+ org.eclipse.egit.ui.internal.staging;version="[6.2.0,7.0.0)",
+ org.eclipse.egit.ui.internal.synchronize.model;version="[6.2.0,7.0.0)",
+ org.eclipse.jgit.errors;version="[6.2.0,7.0.0)",
+ org.eclipse.jgit.lib;version="[6.2.0,7.0.0)",
+ org.eclipse.jgit.revwalk;version="[6.2.0,7.0.0)",
+ org.eclipse.jgit.transport;version="[6.2.0,7.0.0)",
+ org.eclipse.ui.plugin;bundle-version="[3.108.0,4.0.0)"
