+++ mylyn.egit/org.eclipse.mylyn.egit.ui/about.html
