+++ mylyn.egit/org.eclipse.mylyn.egit.ui/src/org/eclipse/mylyn/egit/internal/ui/messages.properties
@@ -1,0 +1,14 @@
+###############################################################################
+# Copyright (c) 2023 Contributors to Eclipse Foundation
+# 
+# This program and the accompanying materials are made available under the
+# terms of the Eclipse Public License v. 2.0 which is available at
+# https://www.eclipse.org/legal/epl-2.0
+# 
+# SPDX-License-Identifier: EPL-2.0
+#
+# Contributors:
+#      See git history
+###############################################################################
+CommitHyperlinkDetector_CommitNotFound=Commit not found
+CommitHyperlinkDetector_CommitNotFoundInRepositories=The commit {0} was not found in any of the configured repositories.
